New income shouldn't be the basis of the alerts
If the cost of an item doesn't violate the current lucky threshold, it shouldn't matter to the player if it violates the new income (as the new income can sometimes be many times larger than the current one, e.g. when buying Reverse Cyclotrons), as they'll still be at least as good as they were before they bought it.