Changed "null" argument to "this" in appendToNative function
Debugging showed that when the Game.tooltip.draw function was called,
the "this" was not what was expected. It was expecting it to be Game.tooltip, but
it was something else. Maybe the Window itself? Adding "this" passed the Game.tooltip
object to the draw function. This seems to have resolved the issue. However, a new tooltip
issue was revealed which will be fixed in my next commit.