Fixes issue #31. Some purchased upgrades not removed from store.
Upgrades like 'Season savings' and 'Santa's dominion' reduce building costs and trigger a storeRefresh
which triggers a simulateBuy before Cookie Clicker code has removed the upgrade from the store. The simulateBuy
resets the upgrade's bought variable back to 0 so it stays in the store when upgrades are rebuilt. Added line
in function isInStore to check for bought=1 so that simulation is prevented.