Fixes issue #50 . Added 2 functions to simulateBuy disable list.
In the latest CC update, the season trigger buyFunction now calls Game.Notify.
I added Game.Notify to the list of function to be disabled in the simulateBuy function.
Also, the Game.Objects['Grandma'].redraw function
was added to the season trigger buyFunction. So we need to add it to the list
of functions that are disabled during a simulateBuy.