Resolves #52 - Stop simulateBuy of Chocolate egg upgrade.
The Chocolate egg was causing the cookie bank value to increase rapidly. Every time
simulateBuy was executed on the Chocolate egg, it would add cookies to the bank, but
the cookies were not removed from the bank afterwards. Since the upgrade does not add Cps, I see no
need to simulate a buy. If the upgrade is later changed to add Cps, then this code will need to be rewritten.